Made a first working CSM implementation. Three variables really matter - smmaxsize which actually gives the size of each cascade - csmsplitn which gives the number of splits - csmfarplane which gives the camera plane after which there is no more shadow. If set to zero, the complete scene is taken into account
The code only supports the shadow2D path and the gatherpcf is not working
since I did not test on a machine supporting it. This is the next thing I am
going to do.

The shading is crappy and incomplete but this should not be a big deal.

The csm code is inactivated now by making the csm insertion fail.

Like Uncharted 2, we should smooth step the transition between
[0, csmfarplane] and ]csmfarplane, +infty[.